// frontend/src/api/generated/gallery/gallery.ts
export type listGalleryDiagramsResponse200 = {
  data: DtoGalleryDiagramSummaryResponse[]
  status: 200
}

export type listGalleryDiagramsResponse500 = {
  data: ListGalleryDiagrams500
  status: 500
}

export type listGalleryDiagramsResponseSuccess = (listGalleryDiagramsResponse200) & {
  headers: Headers;
};
export type listGalleryDiagramsResponseError = (listGalleryDiagramsResponse500) & {
  headers: Headers;
};

export type listGalleryDiagramsResponse = (listGalleryDiagramsResponseSuccess | listGalleryDiagramsResponseError)
// ...
export type getGalleryDiagramResponse200 = {
  data: DtoGalleryDiagramDetailResponse
  status: 200
}

export type getGalleryDiagramResponse404 = {
  data: GetGalleryDiagram404
  status: 404
}

export type getGalleryDiagramResponse500 = {
  data: GetGalleryDiagram500
  status: 500
}

export type getGalleryDiagramResponseSuccess = (getGalleryDiagramResponse200) & {
  headers: Headers;
};
export type getGalleryDiagramResponseError = (getGalleryDiagramResponse404 | getGalleryDiagramResponse500) & {
  headers: Headers;
};

export type getGalleryDiagramResponse = (getGalleryDiagramResponseSuccess | getGalleryDiagramResponseError)